Move ID utilities to a separate file
